refactor(dev-middleware): move server ref with lan URLs to server utilities
