+++ packages/dev-middleware/src/__tests__/InspectorProxyCdpRewritingHacks-test.js
@@ -20,13 +20,13 @@ import {
 } from './InspectorProtocolUtils';
 import {withAbortSignalForEachTest} from './ResourceUtils';
 import {
+  getServerRefWithLanUrls,
   serveStaticJson,
   serveStaticText,
   withServerForEachTest,
-} from './ServerUtils';
+} from "./ServerUtils";
 import {createHash} from 'crypto';
 import fs from 'fs';
-import {networkInterfaces} from 'os';
 import path from 'path';
 
 const fsPromise = fs.promises;
@@ -106,35 +106,8 @@ describe.each(['HTTP', 'HTTPS'])(
         }),
       );
 
-      // Find the current LAN IP address of this machine
-      const lanInterface = Object.values(networkInterfaces())
-        .flat()
-        .find(interface => !interface.internal && interface.family === 'IPv4');
-      if (!lanInterface) {
-        throw new Error('No IPv4 LAN interface found');
-      }
-
-      const serverRefWithLanUrls = {
-        ...serverRef,
-        serverBaseUrl: serverRef.serverBaseUrl.replace(
-          'localhost',
-          lanInterface.address,
-        ),
-        serverBaseWsUrl: serverRef.serverBaseWsUrl.replace(
-          'localhost',
-          lanInterface.address,
-        ),
-      };
-      if (serverRefWithLanUrls.serverBaseUrl === serverRef.serverBaseUrl) {
-        throw new Error(
-          'Failed to replace localhost with LAN address for "serverBaseUrl"',
-        );
-      }
-      if (serverRefWithLanUrls.serverBaseWsUrl === serverRef.serverBaseWsUrl) {
-        throw new Error(
-          'Failed to replace localhost with LAN address for "serverBaseWsUrl"',
-        );
-      }
+      // Get the server ref with LAN addresses instead of localhost
+      const serverRefWithLanUrls = getServerRefWithLanUrls(serverRef);
 
       // Connect to the target using the LAN address instead of localhost
       const {device, debugger_} = await createAndConnectTarget(
